fix: refactor dashboard structure
