+++ src/components/Dashboard/Calendar/WeekEvent/WeekEvent.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import './viewStyles.css';
+import '../styles/viewStyles.css';
 import './WeekEventStyles.css';
 
 const WeekEvent = ({ event }) => {
